Function all_push_cdn::infra::initialize_builder

async fn initialize_builder<TYPES: NodeType<Transaction = TestTransaction, BlockHeader = TestBlockHeader, InstanceState = TestInstanceState>>(
    run_config: &mut NetworkConfig<<TYPES as NodeType>::SignatureKey>,
    validator_config: &ValidatorConfig<<TYPES as NodeType>::SignatureKey>,
    args: &ValidatorArgs,
    orchestrator_client: &OrchestratorClient,
) -> Option<Box<dyn BuilderTask<TYPES>>>
where
    <TYPES as NodeType>::ValidatedState: TestableState<TYPES>,
    <TYPES as NodeType>::BlockPayload: TestableBlock<TYPES>,
    Leaf<TYPES>: TestableLeaf,

Sets correct builder_url and registers a builder with orchestrator if this node is running one. Returns a BuilderTask if this node is going to be running a builder.